+++ db/schema/0016_extend_sessions.up.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `session` ADD COLUMN `extended` UInt16 DEFAULT 0;